requirements.txt
flask
pandas
openpyxl
google-generativeai
python-dotenv
flask-cors
